Data binding for movie list

+++ app/src/main/java/org/sco/movieratings/db/MovieSchema.kt
@@ -15,5 +15,5 @@ data class MovieSchema(
     var overview: String,
     var releaseDate: String,
     var popularity: Double,
-    var voteAverage: Double
+    var voteAverage: Float
 ) : Parcelable
